fix(coord): 放宽 mixing 编码范围校验
- 将 native coordinate 的 mixing 解析扩展到更大的 base36 数值范围
- 同步更新前端校验与搜索规则编辑器的 mixing 输入限制
- 收紧控制侧请求校验，避免旧范围判断误拒合法坐标

+++ src/Setting/NativeCoordinate.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <string>
 
 namespace NativeCoordinate {
@@ -7,7 +8,7 @@ namespace NativeCoordinate {
 struct NativeCoordinateResolution {
     int worldType = -1;
     int seed = 0;
-    int mixing = 0;
+    uint64_t mixing = 0;
     std::string code;
 };
